pyproject.toml
[tool.poetry]
name = "deep-image-prior"
version = "0.1.0"
description = ""
authors = ["Ziyi Xi <ziyixi@mail.ustc.edu.cn>"]
license = "MIT"
readme = "README.md"
packages = [{include = "deep_image_prior"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "1.12.1"
torchvision = "0.13.1"
hydra-core = "^1.2.0"
omegaconf = "^2.2.3"
tensorboard = "^2.11.0"
pytorch-lightning = "^1.8.3.post1"
torchmetrics = "^0.11.0"
matplotlib = "^3.6.2"
segmentation-models-pytorch = "^0.3.1"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.7"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
